Add Email Functionality (#36)
* Added flight API

* Added weather API

* Fixed router to add weather and flight info routes

* Added dependencies for HTTP requests

* Cleaned up router a bit

* Changed flight price currency from EUR to USD

* Created a resource for external API's and updated weather schema

* Starting to store weather data in database

Weather API calls return data from the database if it already exists and makes an external request if none exists

* Fixed forecast dates

* Changed Weather API to check against the database

Checks if a requested city's weather info exists in the database. If it does, then returns that. Otherwise, makes an external request for new data. Also updates outdated weather data in the database.

* Added new field to Weather schema and renamed two existing fields

Added a text field for weather condition and renamed the "high_temp" field to "high" and the "low_temp" field to "low" for more consistency with the forecast fields.

* Created a Flight schema and updated Flight controller functions

* Weather fixes

* Updated Flight schema

* Flight fix

* More fixes to resolve merge conflicts

* Fixed flight database retrieval and updated to use real data

* Made any necessary changes to get the new flights schema to compile. Will need to add flights back to the UI components

* Fixed a typo

* Added a function to retrieve all flights in the database

* Added basic email functionality

* Email fixes